Move TracePatch tests to own file
